Rename picojson_fffd.h to picojson_util.h
